+++ test/manual/watch-template-function/src/app.js
@@ -1,0 +1,9 @@
+import tmpl from './partials/content.html';
+
+const html = tmpl({
+  name: 'World',
+});
+
+document.getElementById('main').innerHTML = html;
+
+console.log('>> app');